Change what coverage reporting reports on by default.
With this change, we change typ -c so that, by default, we only
monitor the python files under the --top-level-dir and any additional
--paths. Note that this will include paths that are *not* imported
during the tests, so that you will be told about files that you missed
completely.

We also add a --coverage-source argument to allow the user to override
the default behavior given above.