Rework typ to handle unimportable modules better.
multiprocessing on windows requires that the __main__ module
be importable in order to work properly. If developers are mostly
working and testing on unix, it's possible for them to not realize this
and attempt to call typ from a wrapper script that may not be
importable, and then get unexpected errors.

This change makes the typ entry points check if it is being called
from an unimportable __main__, and error out if so. Developers
must then specify several different ways to work around this by passing
the 'win_multiprocessing' flag to typ.main() or typ.Runner.main().

The flag takes one of the values specified in the typ.WinMultiprocessing
class:

  ignore -  The tests are not expected to run on Windows, so ignore
            the fact that __main__ is not importable. (If the tests
            are actually run on Windows, it will err out).

  run_serially - Run serially (one at a time) on Windows, regardless
            of how many cores are available (hence multiprocessing is
            not used).

  spawn -   Typ will convert the args it was passed back into a command
            line, and spawn off an invocation of itself on Windows to actually
            execute the tests in parallel (Since typ itself is importable,
            this works fine). This mode cannot be used in conjunction
            with contexts or setup and teardown functions at this time.

  force -   This is like 'spawn', except that it forces the spawn even
            on non-Windows platforms. This mostly exists for testing.

As part of this change, we also no longer expose spawn_main() as a public
entry point (main() handles spawning instead), and merge the cmdline.py
code directly into runner, as it didn't make sense to split things
across the two modules once runner had to do the importability checking.

We also add a host.call_inline() wrapper that is a dumb wrapper
around subprocess.call(); it does not do the output capturing that
host.call() does, and is used when spawning the typ subprocess if needed.