Rework windows multiprocessing approach.
Now we default to spawning unless told not to (what seems likely to be
the common case).

We also rework the way we handle hooks like the classifier and
setup/teardown functions so that they can only be set
from a caller that is importable (and doesn't spawn).

This simplifies the APIs and the implementation.