Remove typ.fakes directory.
It seems that having the fakes in a typ.fakes directory does not
work well with egg files; typ wouldn't launch when run out of an
install.

To work around this until such time as I understand what's going on
better, I've moved the fakes into the top-level directory and
moved the tests for the fakes into typ/tests.