Fix issues arising when load_tests would raise weird exceptions.
This was encountered in https://codereview.chromium.org/838253005/ .

@unittest.skipUnless() was failing in a weird way depending on the
runtime state of the Python process, and so loadTestsFromName() was
raising an exception other than ImportError. typ was not catching
this exception, leading to an unexpected exception from the worker
getting thrown, with little debugging logic.

This patch does two things: first, it catches the exceptions and
the corresponding traceback, fails the test and attaches the error
message. Second, it also captures the traceback when the worker
throws an unexpected exception.

These fixes still need tests; it's not yet obvious to me how to
write a unit test that accurately mimics what was actually going
on in telemetry.