Add a -vvv mode to log when tests are queued for running.
If one is running a bunch of tests in parallel and something
is not working right, it can be useful to see which tests are
currently executing at the same time. There isn't a great way
to do this in typ, because we don't know when tests are actually
picked up for execution by the child processes, but we can at
least log which tests has been queued for running.

This patch adds a -vvv option (verbose=3). This will make
the test log twice as long, since we will now log lines for
when every test starts and stops, but at least that can help
with debugging.