Change typ's interpretation of a test that first fails and is then skipped.
Previously, if a test failed and typ retried it, but for whatever reason
the test was skipped on subsequent retries, the test was considered to
have failed (since it never passed) and so the whole test run failed.

This is perhaps inconsistent with how typ would've reacted if the test
was skipped *every* time, and so arguably this is the wrong thing to do.
For now, I'm changing the behavior so that subsequent skips are treated
like subsequent passes (i.e., the test didn't fail every time).

If you really want different behavior you can look at the full list
of results yourself.

Also, bump the typ version to 0.9.7, and add a codereview.settings flag
so that I can use Rietveld.

R=nednguyen@google.com
BUG=618330