Make multiple top-level dirs work, bump version to 0.10.0.
Previously all of the tests in a given typ invocation had to
exist in a single importable python tree. In other words,
there had to be a single directory at the top and you could
import any module or file from under that directory. This
meant that you couldn't run tests from the
//third_party/WebKit/Tools/Scripts/webkitpy and
//third_party/blink/tools directories of a Chromium checkout
in a single invocation, and that meant you had to jump through
hoops (potentially incorrectly) to try and get coverage in a
single "test suite".

This CL adds support for multiple disjoint top-level directories,
adding a new --top-level-dirs flag that can be set multiple times
(i.e, as a list). The existing --top-level-dir flag still works,
and is equivalent to setting top-level-dirs once. However, setting
both is an error.